Fix weights on sim resampling procedures to use input weight mean
Fix output weights of resampling procedures in downsample_sim and
upsample_sim_inflating_total_weight to use mean of the input sim weights as the
weight for each resampled trajectory.  Prior to this change, the distribution
will be distorted, as the weights were used both as resampling probabilities and
as importance weights.

Change upsample_sim_inflating_total_weight to append the resampled results to
the input sim.