Fix upsample_sim_inflating_total_weight dependencies based on index
Fix upsample_sim_inflating_total_weight so that the outputted ys and weights are
independent of their indices.  The previous change to include all inputted ys
and weights at the beginning of the output ys and output weights followed by
resampling results (if any) produced a dependency.

Add helper function shuffle_sim_indices to perform this index shuffling on a sim
object.