Add array_proxy class, use.proxy argument to reduce memory usage
Add array_proxy class that acts like an array or arraylike object, but reads
elements out off of a filesystem instead of storing them in process memory. For
convenience of implementation, currently stores an integer index instead of the
result, and uses other information to calculate the filename and read from the
appropriate RDS when necessary; this linear process memory overhead could
potentially be avoided with more coding effort, although savings are much easier
if the indexing operations allowed are restricted.

Add the use.proxy option to map_join_ and map_join to have them produce an
array_proxy result instead of a list-mode array holding the results in process
memory.

Use use.proxy=TRUE in generate-retro-and-prospective-forecasts.R to reduce
process memory usage and out-of-memory errors when forking + memory management
causes duplication filling up RAM.