Fix bad weighted quantile calculations when weights sum to 1
Hmisc::wtd.quantile gives undesired results when weights sum to 1, but some
backcasters/forecasters are built to give weights that sum to 1. Pass
normwt=TRUE always to prevent these undesired results in wtd.quantile.or.na, and
remove this as a parameter from the function to prevent confusion about
differing defaults.