Fix multibin score calculation for >1 valid center bins (>1 pkwk)
Fix multibin score calculation when there is greater than one valid center bin,
e.g., when there is greater than one valid peak week in FluSight projects. The
multibin.neighbor.matrix %*% bin.is.valid.center operation should be done using
Boolean AND&OR operations, not numeric; using numeric can cause the resulting
bin.is.accepted to contain entries that are not 0 or 1, e.g., 2, and results in
higher, often invalid (> 0), multibin log scores. Fix this by making the matrix
operation act like using Boolean operations by appending `> 0` to the end.
(Perhaps this issue was caused by a change in behavior of the Matrix package?
Given the above, bin.is.accepted was numeric but was expected to be Boolean
flags in the following line.)