Fix mimicPast* compatibility issues with later dpylr versions
Later versions of dplyr, when group_by & summarize'ing on an empty tibble with a
character factor grouping column, seem to produce either (a) an error about
implicit NA's in the factor column, or (b) a warning and converting the factor
column to an integer column; both break the mimicPast* function.  Explicitly
check for this case and handle it.