Resolve test warnings: better deal with incompatible-length vectors
Tests were using a cbind approach that was combining length-52 and length-53
vectors by recycling the first element of the length-52 vectors to make them
length 53; seems like it could be particularly bad if this happened to a
length-52 test season, but also would introduce discontinuities in training
seasons at the very end. The replacement approach simply copies the last
element. (This old incompatible-length cbind approach seems like it might have
been limited to the tests.)